git-svn-id: svn://192.168.1.10/public/CoffeeMud@8290 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/fakedb/ResultSet.java
@@ -280,7 +280,7 @@ public class ResultSet implements java.sql.ResultSet
    {
 	  if(!showColMap.containsKey(columnName))
 		  return -1;
-	  return showColMap.get(columnName)+1;
+	  return showColMap.get(columnName).intValue()+1;
    }
 
    public String getString(String columnName) throws java.sql.SQLException
